Redirect to issue page after creating a new issue
Previous behavior was to redirect to the issue list with a "successful
creation" message.  This patch will redirect to the page for the
newly-created issue, still with the "successful creation" message.

This matches the behavior after editing an issue and also provides
instant feedback for the user to see if anything went wrong.

Closes #261 and uses the patch contained therein


git-svn-id: https://svn.redmine.org/redmine/trunk@1243 e93f8b46-1217-0410-a6f0-8f06a7374b81